backoffice UI, templated configurable backoffice section implementations (WIP)

+++ openi40.platform/openi40.dbmodel.repositories/src/main/java/com/openi40/dbmodel/repositories/OI40BaseRepository.java
@@ -3,8 +3,10 @@ package com.openi40.dbmodel.repositories;
 import java.sql.Timestamp;
 import java.util.ArrayList;
 import java.util.Date;
+import java.util.HashMap;
 import java.util.Iterator;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 
 import javax.validation.constraints.NotNull;
@@ -28,13 +30,14 @@ import com.openi40.dbmodel.utils.AutoCompleteData;
 import com.openi40.dbmodel.utils.LookupData;
 import com.openi40.dbmodel.utils.PageInfo;
 import com.openi40.dbmodel.utils.QbeSupport;
+
 /**
  * 
  * This code is part of the OpenI40 open source advanced production scheduler
- * platform suite, have look to its licencing options.
- * Web site: http://openi40.org/  
- * Github: https://github.com/openi40/OpenI40Platform
- * We hope you enjoy implementing new amazing projects with it.
+ * platform suite, have look to its licencing options. Web site:
+ * http://openi40.org/ Github: https://github.com/openi40/OpenI40Platform We
+ * hope you enjoy implementing new amazing projects with it.
+ * 
  * @author architectures@openi40.org
  *
  */
@@ -49,19 +52,34 @@ public interface OI40BaseRepository<OI40Type extends OI40DBBaseEntity> extends J
 			return null;
 	}
 
-	@GetMapping(path = "deleteByCode")
-	@Transactional
-	default public void deleteByCode(@NotNull @RequestParam("code") String code) {
-		this.deleteById(code);
+	@GetMapping(path = "isDeletable", produces = MediaType.APPLICATION_JSON_VALUE)
+	default public boolean isDeletable(@NotNull @RequestParam("code") String code) {
+		return false;
 	}
 
-	@PostMapping(path = "deleteByCodes")
+	@GetMapping(path = "deleteByCode", produces = MediaType.APPLICATION_JSON_VALUE)
 	@Transactional
-	default public void deleteByCodes(@RequestBody List<String> codes) {
+	default public boolean deleteByCode(@NotNull @RequestParam("code") String code) {
+		if (isDeletable(code)) {
+			this.deleteById(code);
+			return true;
+		}
+		return false;
+	}
+
+	@PostMapping(path = "deleteByCodes", produces = MediaType.APPLICATION_JSON_VALUE)
+	@Transactional
+	default public Map<String,Boolean> deleteByCodes(@RequestBody List<String> codes) {
+		Map<String,Boolean> retValue=new HashMap<String, Boolean>();
 		for (Iterator iterator = codes.iterator(); iterator.hasNext();) {
 			String code = (String) iterator.next();
-			this.deleteById(code);
+			boolean canBeDeleted=isDeletable(code);
+			if (canBeDeleted) {
+				this.deleteById(code);
+			}
+			retValue.put(code, canBeDeleted);
 		}
+		return retValue;
 	}
 
 	@PostMapping(path = "findByCodes", produces = MediaType.APPLICATION_JSON_VALUE)
@@ -89,13 +107,15 @@ public interface OI40BaseRepository<OI40Type extends OI40DBBaseEntity> extends J
 		}
 		return outdata;
 	}
+
 	@PostMapping(path = "updateSingle", produces = MediaType.APPLICATION_JSON_VALUE)
 	@Transactional
 	default public OI40Type updateSingle(@RequestBody OI40Type data) {
 		data.setModifiedTimestamp(new Date());
-		data = this.saveAndFlush(data);		
+		data = this.saveAndFlush(data);
 		return data;
 	}
+
 	@GetMapping(produces = MediaType.APPLICATION_JSON_VALUE)
 	List<OI40Type> findAll();
 
@@ -112,7 +132,7 @@ public interface OI40BaseRepository<OI40Type extends OI40DBBaseEntity> extends J
 
 	@PostMapping(path = "findByQbePaged", produces = MediaType.APPLICATION_JSON_VALUE, consumes = MediaType.APPLICATION_JSON_VALUE)
 	default public Page<OI40Type> findByQbePaged(@NotNull @RequestBody QbeSupport<OI40Type> qbe) {
-		Pageable p=PageInfo.from(qbe.getPage());
+		Pageable p = PageInfo.from(qbe.getPage());
 		Page<OI40Type> out = this.findAll(Example.of(qbe.getQbe()), p);
 		return out;
 	}
@@ -124,17 +144,29 @@ public interface OI40BaseRepository<OI40Type extends OI40DBBaseEntity> extends J
 	@GetMapping(path = "modifiedAfter/{ts}")
 	@Query("select e from #{#entityName} e where e.modifiedTimestamp>=?1")
 	public List<OI40Type> findByAfterModifiedTimestamp(@NotNull @PathVariable("ts") Timestamp ts);
-	
-	//@Query("select e from #{#entityName} e where ((?1 is null || ?1=='')  OR  (upper(e.code) LIKE (upper(?1) || '%' ) )) AND ((?2 is null || ?2=='')  OR  (upper(e.description) LIKE ('%' || upper(?2) || '%' ) ))")
-	public Page<OI40Type> findByCodeStartsWithIgnoreCaseAndDescriptionContainsIgnoreCase(String code,String description,Pageable page) ;
-	public Page<OI40Type> findByCodeContainsIgnoreCaseOrDescriptionContainsIgnoreCase(String code,String description,Pageable page) ;
+
+	// @Query("select e from #{#entityName} e where ((?1 is null || ?1=='') OR
+	// (upper(e.code) LIKE (upper(?1) || '%' ) )) AND ((?2 is null || ?2=='') OR
+	// (upper(e.description) LIKE ('%' || upper(?2) || '%' ) ))")
+	public Page<OI40Type> findByCodeStartsWithIgnoreCaseAndDescriptionContainsIgnoreCase(String code,
+			String description, Pageable page);
+
+	public Page<OI40Type> findByCodeContainsIgnoreCaseOrDescriptionContainsIgnoreCase(String code, String description,
+			Pageable page);
+
 	@PostMapping("doAutocomplete")
 	default public Page<OI40Type> doAutocomplete(@NotNull @RequestBody AutoCompleteData autoCompleteData) {
-		return this.findByCodeContainsIgnoreCaseOrDescriptionContainsIgnoreCase(autoCompleteData.getSearchString()!=null?autoCompleteData.getSearchString().trim():"",autoCompleteData.getSearchString()!=null?autoCompleteData.getSearchString().trim():"", PageInfo.from(autoCompleteData.getPage()));
+		return this.findByCodeContainsIgnoreCaseOrDescriptionContainsIgnoreCase(
+				autoCompleteData.getSearchString() != null ? autoCompleteData.getSearchString().trim() : "",
+				autoCompleteData.getSearchString() != null ? autoCompleteData.getSearchString().trim() : "",
+				PageInfo.from(autoCompleteData.getPage()));
 	}
+
 	@PostMapping("doLookup")
 	default public Page<OI40Type> doLookup(@NotNull @RequestBody LookupData lookup) {
-		return this.findByCodeStartsWithIgnoreCaseAndDescriptionContainsIgnoreCase(lookup.getCode()!=null?lookup.getCode().trim():"",lookup.getDescription()!=null?lookup.getDescription().trim():"", PageInfo.from(lookup.getPage()));
+		return this.findByCodeStartsWithIgnoreCaseAndDescriptionContainsIgnoreCase(
+				lookup.getCode() != null ? lookup.getCode().trim() : "",
+				lookup.getDescription() != null ? lookup.getDescription().trim() : "", PageInfo.from(lookup.getPage()));
 	}
 
 }
